requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
psycopg2-binary==2.9.9
kubernetes==28.1.0
python-multipart==0.0.6
boto3==1.34.0
google-cloud-storage==2.14.0
azure-storage-blob==12.19.0

